:white_check_mark: test: add focused git mutation tests
